subindo para a vercel

+++ src/components/TableProductsConfig/index.tsx
@@ -1,38 +1,39 @@
-import React, { useEffect, useState } from 'react';
+import React, { Dispatch, SetStateAction, useCallback, useEffect, useState } from 'react';
 import { ButtonDel, ButtonUpdate, Container, Table, TableBody, TableData, TableHeader, TableRow, TitleColumn, } from "./styles";
 import api from "../../services/api"
 import { format } from "date-fns"
 import { useUpdateProduct } from '../../context/updateProductContext';
+import { IProducts } from '../../pages/DashBoard';
+import { useToasts } from '@geist-ui/react';
 
 
 
-interface ITableProductsProps {
-
+interface ITableProductsProps{
+  products: IProducts[],
+  setProducts: Dispatch<SetStateAction<IProducts[]>>,
   onOpenModalUpdate: () => void
 
   }
 
-  interface IProducts {
-    id: string;
-    name_product: string;
-    obs_product: string;
-    quantity: string;
-    created_at: string;
-    updated_at: string;
-  }
+  
 
 
-export function TableProductsConfig({ onOpenModalUpdate } : ITableProductsProps) {
+export function TableProductsConfig({ products, setProducts, onOpenModalUpdate } : ITableProductsProps) {
 
   
-  const [products, setProducts] = useState<IProducts[]>([]);
-  const { id, setId } = useUpdateProduct();
-  const { id_creator, setIdCreator } =useUpdateProduct();
-  const { name_products, setNameProduct } = useUpdateProduct();
-  const { obs_products, setObsProduct } = useUpdateProduct();
-  const { quantity, setQuantity } = useUpdateProduct();
-  const { created_at, setCreatedAt } = useUpdateProduct();
-  const { update_at, setUpdateAt } = useUpdateProduct();
+  
+  const { setId } = useUpdateProduct();
+  const { id_creator, setIdCreator } = useUpdateProduct();
+  const { setNameProduct } = useUpdateProduct();
+  const { setObsProduct } = useUpdateProduct();
+  const { setQuantity } = useUpdateProduct();
+
+  const [, setToast] = useToasts();
+
+  const userLoggedString = localStorage.getItem("user_logged")
+  const currentData = userLoggedString ? JSON.parse(userLoggedString) : []
+
+  
  
 
   function editProduct(id: string, name_products: string, obs_products: string, quantity: string ) {
@@ -40,17 +41,17 @@ export function TableProductsConfig({ onOpenModalUpdate } : ITableProductsProps)
     setNameProduct(name_products)
     setObsProduct(obs_products)
     setQuantity(quantity)
-    // setId(id);
-    // setNameProduct(name_products);
-    // setObsProduct(obs_products);
-    // setQuantity(quantity);
-    console.log('oi');
-    
-    
   }
 
   async function deleteProduct(id: string) {
     try {
+
+      if (currentData.admin === false) {
+        return setToast({
+          text: 'Somente Administradores podem deletar produtos!.',
+          type: 'error'
+        })
+      }
       await api.delete(`products/${id}`);
       
       setProducts((oldProducts) =>
@@ -61,19 +62,8 @@ export function TableProductsConfig({ onOpenModalUpdate } : ITableProductsProps)
     }
   }
 
-
-  useEffect(() => {
-    api.get(`products`).then((response) => {
-
-      setProducts(response.data);
-
-    }); 
-  }, [products]);
-    
     return (
       <Container>
-        
-       
         <Table>
           <TableHeader>
             <TableRow>
@@ -89,23 +79,23 @@ export function TableProductsConfig({ onOpenModalUpdate } : ITableProductsProps)
           <TableBody>
           
           {products.map((product) => ( 
-          <TableRow key={product.id}>
-            <TableData> {product.name_product} </TableData>
-            <TableData> {product.obs_product} </TableData>
-            <TableData> {product.quantity} </TableData>
-            <TableData> {format( new Date (product.created_at), "dd/MM/yyyy HH:mm:ss")} </TableData>
-            <TableData> {format( new Date (product.updated_at), "dd/MM/yyyy HH:mm:ss")} </TableData>
-            <TableData> 
-              <ButtonUpdate type="submit" onClick={() => { onOpenModalUpdate(); editProduct(product.id, product.name_product, product.obs_product, product.quantity  )}} >
-                Editar
-              </ButtonUpdate>
-            </TableData>
-            <TableData> 
-              <ButtonDel type="submit" id = "btnDelet" onClick={() =>{deleteProduct(product.id) }}>
-                Deletar
-              </ButtonDel> 
-            </TableData>
-          </TableRow>
+            <TableRow key={product.id}>
+              <TableData> {product.name_product} </TableData>
+              <TableData> {product.obs_product} </TableData>
+              <TableData> {product.quantity} </TableData>
+              <TableData> {format( new Date (product.created_at), "dd/MM/yyyy HH:mm:ss")} </TableData>
+              <TableData> {format( new Date (product.updated_at), "dd/MM/yyyy HH:mm:ss")} </TableData>
+              <TableData> 
+                <ButtonUpdate type="submit" onClick={() => { onOpenModalUpdate(); editProduct(product.id, product.name_product, product.obs_product, product.quantity  )}} >
+                  Editar
+                </ButtonUpdate>
+              </TableData>
+              <TableData> 
+                <ButtonDel type="submit" id = "btnDelet" onClick={() =>{deleteProduct(product.id) }}>
+                  Deletar
+                </ButtonDel> 
+              </TableData>
+            </TableRow>
           ))}
                
         </TableBody>
